+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-06-11 12:33:07
-// Commit: feat(network): integrate custom telemetry logger in network
+// Updated at: 2026-06-11 22:54:41
+// Commit: fix(network): fix memory leak by cancelling network subscription
 
 class networkComponentState {
-  final String commitRef = '1898';
-  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
-  final DateTime timestamp = DateTime.parse('2026-06-11T12:33:07+05:00');
+  final String commitRef = '1928';
+  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-11T22:54:41+05:00');
 }